Ok so this is super dangerous.
In all our schemas (and the option case) we have our recursive branch to the left e.g. Option[A] ~= A \/ Unit. With our Person schema recursing via Option[Role] we had the recursive case in the left. In Play Json Reads we handled coproducts by trying left before right ending in an infinite expansion of the recursive case because the terminal case never got tried.